Refactor the code.
Slightly refactores the code, improving the overall usability
and readibility, as well as maintainability.

The changes:

1. As long as this is a project for just one game, and the game rules
mandate having only two players, not more or less, it is better to have
a more precise code that can only operate with two objects, instead of
a more generic integer which can contain any other values and will
require constant checks everywhere. It is impossible to set any other
player within the code.
2. Same goes for the winner - it is either one or another player wins,
or there is a tie. This commit makes the code much more simple, strict
and deterministic. It is impossible to set any other winner.
3. Refactors the Controller's response bodies to return a class, instead
of a map object, for better readibility and agility: now it is always
known which fields there can be and which not, how exactly those will be
named. The user code (for example, the UI), can also count on that.
4. Adds more unit tests.
5. For better readability and maintainability, the code now uses Stream
objects when it is possible or makes sense.